Add config file support
It can be more convenient to have all the parameters set in config
files rather than passing them to the service wrapper.

Also, at the moment, it's impossible to pass a command to be
executed before running the service, containing whitespaces in the
binary path (e.g. --exec-start-pre="'some binary' 'some arg'").